+++ addon/index.js
@@ -1,7 +1,0 @@
-import memberAction from './utils/member-action';
-import collectionAction from './utils/collection-action';
-
-export const classOp = collectionAction;
-export const instanceOp = memberAction;
-
-export { collectionAction, memberAction };