+++ tests/dummy/.eslintrc.js
@@ -1,0 +1,14 @@
+/* eslint-env node */
+module.exports = {
+  root: true,
+  parserOptions: {
+    ecmaVersion: 2017,
+    sourceType: 'module'
+  },
+  plugins: ['ember'],
+  extends: ['eslint:recommended', 'plugin:ember/recommended'],
+  env: {
+    browser: true
+  },
+  rules: {}
+};