+++ tests/dummy/app/controllers/index.js
@@ -1,8 +1,9 @@
-import Ember from 'ember';
-
-const { Controller, A } = Ember;
+import { A } from '@ember/array';
+import Controller from '@ember/controller';
+import Fruit from '../models/fruit';
 
 export default Controller.extend({
+  requests: [],
   init() {
     this._super(...arguments);
     this.set('requests', A([]));
@@ -10,12 +11,10 @@ export default Controller.extend({
   // BEGIN-SNIPPET controller
   actions: {
     ripenFruit(fruit) {
-      fruit.ripen(
-        fruit.getProperties(['id', 'name'])
-      );
+      fruit.ripen(fruit.getProperties(['id', 'name']));
     },
     fruitInfo(fruit) {
-      let { id } = fruit.getProperties(['id', 'name']);
+      const { id } = fruit.getProperties(['id', 'name']);
       fruit.info({ fruitId: id });
     },
     ripenAllFruit(fruit) {
@@ -23,6 +22,15 @@ export default Controller.extend({
     },
     getAllFreshFruit(fruit) {
       fruit.getFresh({ month: 'July' });
+    },
+    eatFruit(fruit) {
+      fruit.eat({ was_eaten: true });
+    },
+    eatAll(fruit) {
+      fruit.eatAll({ was_eaten: true });
+    },
+    juiceAllFruit(fruit) {
+      Fruit.juiceAll({ was_eaten: true });
     }
   }
   // END-SNIPPET