+++ tests/dummy/app/helpers/first-in-array.ts
@@ -1,8 +1,10 @@
+import { isArray } from '@ember/array';
+import Helper from '@ember/component/helper';
 import Ember from 'ember';
 
-const { isArray, Helper, HTMLBars } = Ember;
+const { HTMLBars } = Ember as any;
 
-export function firstInArray(params/* , hash*/) {
+export function firstInArray(params: any[] /* , hash*/) {
   if (isArray(params[0])) {
     return params[0].objectAt(0) || undefined;
   } else {