+++ tests/dummy/app/helpers/or.ts
@@ -1,0 +1,18 @@
+import Helper from '@ember/component/helper';
+import Ember from 'ember';
+
+const { HTMLBars } = Ember as any;
+
+export function or([a, b]: any[] /* , hash*/) {
+  return a || b;
+}
+
+let forExport = null;
+
+if (Helper) {
+  forExport = Helper.helper(or);
+} else if (HTMLBars.makeBoundHelper) {
+  forExport = HTMLBars.makeBoundHelper(or);
+}
+
+export default forExport;