+++ tests/helpers/start-app.js
@@ -1,12 +1,12 @@
-import Ember from 'ember';
 import Application from '../../app';
 import config from '../../config/environment';
-
-const { merge, run } = Ember;
+import { assign } from '@ember/polyfills';
+import { run } from '@ember/runloop';
 
 export default function startApp(attrs) {
-  let attributes = merge({}, config.APP);
-  attributes = merge(attributes, attrs); // use defaults, but you can override;
+  let attributes = assign({}, config.APP);
+  attributes.autoboot = true;
+  attributes = assign(attributes, attrs); // use defaults, but you can override;
 
   return run(() => {
     let application = Application.create(attributes);